Augmented-Dickey-Fuller Unit Root Test

Performs the augmented Dickey-Fuller unit root test in time series, testing whether a time series is stationary. It extends the simple Dickey-Fuller test by including lagged difference terms to account for autocorrelation.

adfTest(
  y,
  type = c("none", "drift", "trend"),
  lags = 1,
  selectlags = c("Fixed", "AIC", "BIC")
)

Arguments

y

Vector to be tested for a unit root.

type

Test type, either "none", "drift" or "trend".

lags

Number of lags for endogenous variable to be included.

selectlags

Lag selection can be achieved according to the Akaike "AIC" or the Bayes "BIC" information criteria. The maximum number of lags considered is set by lags. The default is to use a "fixed" lag length set by lags.

Value

An object of class htest.

Details

The function adfTest() computes the augmented Dickey-Fuller test. If type is set to "none" neither an intercept nor a trend is included in the test regression. If it is set to "drift" an intercept is added and if it is set to "trend" both an intercept and a trend is added. The critical values are taken from Hamilton (1994) and Dickey and Fuller(1981).

Note

Adapted from code by Bernhard Pfaff to conform to package standards.

References

Dickey, D. A. and Fuller, W. A. (1979), Distributions of the Estimators For Autoregressive Time Series with a Unit Root, Journal of the American Statistical Association, 75, 427–431.

Dickey, D. A. and Fuller, W. A. (1981), Likelihood Ratio Statistics for Autoregressive Time Series with a Unit Root, Econometrica, 49, 1057–1072.

Hamilton (1994), Time Series Analysis, Princeton University Press.

See also

Examples

lc.df <- adfTest(y=AirPassengers, lags=3, type='trend')
summary(lc.df)
#>                 Length Class  Mode     
#> statistic        3     -none- numeric  
#> parameter        1     -none- numeric  
#> critical.values  9     -none- numeric  
#> method           1     -none- character
#> data.name       38     -none- character