Dirichlet distribution functions

Density, random generation, and basic utilities for the Dirichlet distribution.

Fast parallel approximation of the Dirichlet CDF using RcppParallel.

Generates random draws from a Dirichlet distribution.

The Dirichlet distribution has no unique multivariate quantile function.

ddirichlet(x, concentration, log = FALSE)

pdirichlet(q, concentration, R = 100000)

rdirichlet(n, concentration)

qdirichlet()

Arguments

x

numeric vector or matrix (rows sum to 1)

concentration

numeric vector of concentration parameters (> 0)

log

logical; return log-density if TRUE

q

numeric vector

R

number of simulations

n

number of samples

Value

Numeric vector of densities

Approximate probability

Matrix with n rows

Details

The Dirichlet distribution is a multivariate generalization of the Beta distribution defined on the simplex: $$\sum_{i=1}^k x_i = 1, \quad x_i \ge 0$$

See also

Examples

ddirichlet(c(0.2, 0.3, 0.5), c(1,1,1))
#> [1] 2

pdirichlet(c(0.2,0.3,0.5), c(1,1,1))
#> [1] 0

rdirichlet(5, c(1,1,1))
#>            [,1]      [,2]      [,3]
#> [1,] 0.10717260 0.7727606 0.1200668
#> [2,] 0.02509533 0.7624872 0.2124175
#> [3,] 0.04689484 0.1765656 0.7765396
#> [4,] 0.53266996 0.3078648 0.1594653
#> [5,] 0.22024252 0.2884604 0.4912971