Kwiatkowski-Phillips-Schmidt-Shin Test for Stationarity in Time Series

A test for stationarity in time series (Kwiatkowski-Phillips-Schmidt-Shin), complementary to unit root tests such as the ADF test. It tests the null hypothesis of stationarity rather than the null of a unit root.

kpssTest(
  y,
  type = c("mu", "tau"),
  lags = c("short", "long", "nil"),
  useLag = NULL
)

Arguments

y

Vector to be tested for a unit root.

type

Type of deterministic part.

lags

Maximum number of lags used for error term correction.

useLag

User specified number of lags.

Value

An object of class htest. In addition to the standard fields, the element critical.values contains the critical values at 10\

Details

Performs the KPSS unit root test, where the Null hypothesis is stationarity. The test types specify as deterministic component either a constant "mu" or a constant with linear trend "tau".

lags="short" sets the number of lags to \(\sqrt[4]{4 \times (n/100)}\), whereas lags="long" sets the number of lags to \(\sqrt[4]{12 \times (n/100)}\). If lags="nil" is choosen, then no error correction is made. Furthermore, one can specify a different number of maximum lags by setting useLag accordingly.

Note

Based on code by Bernhard Pfaff

References

Kwiatkowski, D., Phillips, P.C.B., Schmidt, P. and Shin, Y., (1992), Testing the Null Hypothesis of Stationarity Against the Alternative of a Unit Root: How Sure Are We That Economic Time Series Have a Unit Root?, Journal of Econometrics, 54, 159–178.

Download possible at: https://cowles.yale.edu/, see rubric 'Discussion Papers (CFDPs)'.

See also

Examples

kpss.gnp <- kpssTest(AirPassengers, type="tau", lags="short")
summary(kpss.gnp)
#>                 Length Class  Mode     
#> statistic        1     -none- numeric  
#> parameter        1     -none- numeric  
#> critical.values  4     -none- numeric  
#> method           1     -none- character
#> data.name       12     -none- character